--- src/test/java/seleniumTestNG/SelectMenu.java
@@ -1,20 +1,0 @@
-package seleniumTestNG;
-
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
-
-public class SelectMenu extends BaseSetUp {
-
-
-    public static Boolean HandlingSelectMenu() {
-        Boolean bol=false;
-        WebElement Widgets_Frames=driver.findElement(By.xpath( "//*[@class=\"category-cards\"]//following::div[@class=\"card-body\"]//h5[contains(text(),\"Widgets\")]"));
-        Widgets_Frames.click();
-        WebElement SelectMenu_span=driver.findElement(By.xpath( "//*[@class=\"accordion\"]//div[@class=\"element-group\"]//following::span[@class=\"text\" and contains(text(),\"Select Menu\")]"));
-        SelectMenu_span.click();
-        bol=true;
-        return bol;
-
-    }
-}
